# specs/001-provider-user-subscriptions/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Provider User Subscriptions API
  version: 1.0.0
  description: >-
    NestJS in-memory API for Company, User, and company-level Subscription management.
servers:
  - url: http://localhost:3000
paths:
  /companies:
    get:
      summary: List companies (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: Paginated company list
    post:
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyDto'
      responses:
        '201': { description: Company created }
        '400': { description: Validation error }
  /companies/{companyId}:
    get:
      summary: Get company by id
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Company found }
        '404': { description: Not found }
    patch:
      summary: Update company
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyDto'
      responses:
        '200': { description: Company updated }
        '400': { description: Validation error }
        '404': { description: Not found }
    delete:
      summary: Delete company and cascade users/subscriptions
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }

  /users:
    get:
      summary: List users (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - in: query
          name: companyId
          schema: { type: string }
      responses:
        '200': { description: Paginated user list }
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201': { description: User created }
        '400': { description: Validation error }
        '404': { description: Company not found }
  /users/{userId}:
    get:
      summary: Get user by id
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200': { description: User found }
        '404': { description: Not found }
    patch:
      summary: Update user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200': { description: User updated }
        '400': { description: Validation error }
        '404': { description: Not found }
    delete:
      summary: Delete user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }

  /subscriptions/{companyId}:
    get:
      summary: Get latest subscription details for company
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Subscription details }
        '404': { description: Company or subscription not found }

  /subscriptions/{companyId}/buy:
    post:
      summary: Buy subscription for company (admin-only via actorUserId)
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuySubscriptionDto'
      responses:
        '201': { description: Subscription purchased }
        '400': { description: Validation or plan/business-rule error }
        '403': { description: Actor not admin or wrong company }
        '404': { description: Company or actor not found }
        '409': { description: Active subscription already exists }

  /subscriptions/{companyId}/cancel:
    post:
      summary: Cancel active subscription for company (admin-only via actorUserId)
      parameters:
        - in: path
          name: companyId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSubscriptionDto'
      responses:
        '200': { description: Subscription canceled }
        '400': { description: Validation or business-rule error }
        '403': { description: Actor not admin or wrong company }
        '404': { description: Company or actor not found }

components:
  schemas:
    CreateCompanyDto:
      type: object
      required: [name, legalName]
      properties:
        name: { type: string }
        legalName: { type: string }
        websiteUrl: { type: string, format: uri }
        email: { type: string, format: email }
        phone: { type: string }
        address: { type: string }
    UpdateCompanyDto:
      type: object
      properties:
        name: { type: string }
        legalName: { type: string }
        websiteUrl: { type: string, format: uri }
        email: { type: string, format: email }
        phone: { type: string }
        address: { type: string }
    CreateUserDto:
      type: object
      required: [companyId, name, email, isAdmin]
      properties:
        companyId: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        isAdmin: { type: boolean }
        phone: { type: string }
        title: { type: string }
    UpdateUserDto:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        isAdmin: { type: boolean }
        phone: { type: string }
        title: { type: string }
    BuySubscriptionDto:
      type: object
      required: [actorUserId, planCode]
      properties:
        actorUserId: { type: string }
        planCode:
          type: string
          enum: [BASIC_MONTHLY, BASIC_YEARLY, STANDARD_MONTHLY, STANDARD_YEARLY]
    CancelSubscriptionDto:
      type: object
      required: [actorUserId]
      properties:
        actorUserId: { type: string }